% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{abs.cg.node}
\alias{abs.cg.node}
\title{Absolute Value}
\usage{
\method{abs}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{abs(x)}.
}
\author{
Ron Triepels
}
