% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{abs}
\alias{abs}
\title{Absolute Value}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{abs(x)}.
}
\section{Usage}{

\preformatted{abs(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
