% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{acos.cg.node}
\alias{acos.cg.node}
\title{Inverse Cosine}
\usage{
\method{acos}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{acos(x)}.
}
\author{
Ron Triepels
}
