% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{acosh.cg.node}
\alias{acosh.cg.node}
\title{Inverse Hyperbolic Cosinus}
\usage{
\method{acosh}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{acosh(x)}.
}
\author{
Ron Triepels
}
