% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{acosh.grad}
\alias{acosh.grad}
\title{Inverse Hyperbolic Cosine Gradient}
\usage{
\method{acosh}{grad}(x, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{acosh(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
