% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{add.grad.x}
\alias{add.grad.x}
\title{Add Gradient}
\usage{
add.grad.x(x, y, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x + y} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
