% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{address}
\alias{address}
\title{Retrieve Object Address}
\usage{
address(x)
}
\arguments{
\item{x, }{any R object, the object whose pointer address is retrieved.}
}
\value{
character scalar, the address of the object.
}
\description{
Retrieve the pointer address of \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
