% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{adj.mat}
\alias{adj.mat}
\title{Adjacency Matrix}
\value{
numeric matrix, the adjacency matrix of the graph.
}
\description{
Create an adjacency matrix of the computational graph.
}
\section{Usage}{

\preformatted{adj.mat(value, name)}
}

\author{
Ron Triepels
}
