% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{array0}
\alias{array0}
\title{Create Numeric Array}
\usage{
array0(x, dim = length(x), dimnames = NULL)
}
\arguments{
\item{x, }{numeric vector or array, the data to fill the array.}

\item{dim, }{integer vector, the dimensions of the array. Defaults to the length of \code{x}.}

\item{dimnames, }{list, names for the dimensions. Is ignored when \code{dimnames} is \code{NULL}. Defaults to \code{NULL}.}
}
\value{
numeric array, \code{x} coerced to an array with dimensions \code{dim}.
}
\description{
Create an numeric array without recycling. Missing values are replaced with zeroes.
}
\note{
In contrast to base function \link[base]{array}, this function only works for numerical inputs and does not recycle elements.
}
\author{
Ron Triepels
}
