% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as.array.cg.node}
\alias{as.array.cg.node}
\title{Coerce to an Array}
\usage{
\method{as.array}{cg.node}(x, dims, name = cgraph::name(), ...)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{dims}{numeric scalar or vector, the dimensions of the new array.}

\item{name}{character scalar, name of the operation (optional).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
cg.node, node of the operation.
}
\description{
Coerce \code{x} to an array with dimensions \code{dims}.
}
\note{
This function is equivalent to function \code{cg.reshape}.
}
\author{
Ron Triepels
}
