% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{asin.cg.node}
\alias{asin.cg.node}
\title{Inverse Sine}
\usage{
\method{asin}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{asin(x)}.
}
\author{
Ron Triepels
}
