% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{asinh.cg.node}
\alias{asinh.cg.node}
\title{Inverse Hyperbolic Sine}
\usage{
\method{asinh}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{asinh(x)}.
}
\author{
Ron Triepels
}
