% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{atan.cg.node}
\alias{atan.cg.node}
\title{Inverse Tangent}
\usage{
\method{atan}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{atan(x)}.
}
\author{
Ron Triepels
}
