% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{atanh.cg.node}
\alias{atanh.cg.node}
\title{Inverse Hyperbolic Tangent}
\usage{
\method{atanh}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{atanh(x)}.
}
\author{
Ron Triepels
}
