% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{c}
\alias{c}
\title{Transform Array to Vector}
\value{
symbol, name of the operation.
}
\description{
Transform array \code{x} to a one-dimensional vector.
}
\section{Usage}{

\preformatted{c(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
