% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.abs.grad}
\alias{cg.abs.grad}
\title{Absolute Value Gradient}
\usage{
cg.abs.grad(x, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{abs(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
