% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.abs}
\alias{cg.abs}
\title{Absolute Value}
\usage{
cg.abs(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{abs(x)}.
}
\author{
Ron Triepels
}
