% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.add.parms}
\alias{cg.add.parms}
\title{Add Parameters}
\arguments{
\item{parms}{named list, the parameters that are to be added to the graph.}
}
\value{
nothing.
}
\description{
Add a list of parameters with default values to the graph.
}
\details{
\code{$add.parms(parms)}
}
\note{
The names of the list elements are taken to be the parameter names.
}
\author{
Ron Triepels
}
