% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cg.as.double.grad}
\alias{cg.as.double.grad}
\title{Coerce to a Numeric Vector Gradient}
\usage{
cg.as.double.grad(x, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{as.numeric(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
