% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cg.as.double}
\alias{cg.as.double}
\title{Coerce to a Numeric Vector}
\usage{
cg.as.double(x, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Coerce \code{x} to a one-dimensional numeric vector.
}
\note{
This function is identical to \code{cg.as.numeric}.
}
\author{
Ron Triepels
}
