% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cg.asin}
\alias{cg.asin}
\title{Inverse Sinus}
\usage{
cg.asin(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{asin(x)}.
}
\author{
Ron Triepels
}
