% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cg.asinh}
\alias{cg.asinh}
\title{Inverse Hyperbolic Sine}
\usage{
cg.asinh(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{asinh(x)}.
}
\author{
Ron Triepels
}
