% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cg.atanh}
\alias{cg.atanh}
\title{Inverse Hyperbolic Tangent}
\usage{
cg.atanh(x, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{atanh(x)}.
}
\author{
Ron Triepels
}
