% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cg.c}
\alias{cg.c}
\title{Transform Array to Vector}
\usage{
cg.c(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Transform array \code{x} to a one-dimensional vector.
}
\author{
Ron Triepels
}
