% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.colSums}
\alias{cg.colSums}
\title{Column Sums}
\usage{
cg.colSums(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{colSums(x)}.
}
\author{
Ron Triepels
}
