% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.const}
\alias{cg.const}
\title{Add Constant}
\arguments{
\item{value}{numeric scalar or array, default value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, an auto-generated name is assigned to the node.}
}
\value{
cg.node, constant node.
}
\description{
Add a constant node to the graph.
}
\details{
\code{$const(value, name)}
}
\note{
Constant nodes are ignored when differentiating a graph.

In case \code{value} has no dimension (i.e. it has no attribute \code{dim} attached to it), the dimension of \code{value} is automatically set to the length of the vector.
}
\author{
Ron Triepels
}
