% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.count.type}
\alias{cg.count.type}
\title{Count Nodes}
\arguments{
\item{type}{numeric scalar, type of the node. Should be either: 0 (constant), 1 (input), 2 (parameter), or 3 (expression). Defaults to 3 (expression).}
}
\value{
numeric scalar, the numer of nodes of the given type.
}
\description{
Count how many nodes of a specific type are added to the graph.
}
\details{
\code{$count.type(type = 3)}
}
\author{
Ron Triepels
}
