% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.crossprod}
\alias{cg.crossprod}
\title{Matrix Crossproduct}
\usage{
cg.crossprod(x, y, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{y}{cg.node, placeholder for a numeric matrix.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{crossprod(x, y)}.
}
\note{
In contrast to the original \code{crossprod} function, this function requires both \code{x} and \code{y} to be supplied.
}
\author{
Ron Triepels
}
