% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.div}
\alias{cg.div}
\title{Divide}
\usage{
cg.div(x, y, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}

\item{y}{cg.node, placeholder for a numeric scalar or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x / y}.
}
\author{
Ron Triepels
}
