% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{cg.equal}
\alias{cg.equal}
\title{Equal}
\usage{
cg.equal(x, y, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x == y}.
}
\author{
Ron Triepels
}
