% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.exp}
\alias{cg.exp}
\title{Exponential Function}
\usage{
cg.exp(x, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{exp(x)}.
}
\author{
Ron Triepels
}
