% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.get.parms}
\alias{cg.get.parms}
\title{Get Parameters}
\value{
named list, parameters of the graph.
}
\description{
List all parameters and their values.
}
\details{
\code{$get.parms()}
}
\author{
Ron Triepels
}
