% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.initialize}
\alias{cg.initialize}
\title{Computational Graph}
\value{
cgraph object.
}
\description{
Initialize a computational graph.
}
\details{
\code{$new()}
}
\note{
The cgraph object is set to be the active graph. Any nodes that are created by wrapper function \link[cgraph]{const}, \link[cgraph]{input}, \link[cgraph]{parm}, or \link[cgraph]{opr} will be added to this graph. Also, when printing a node, its value will be evaluated in the active graph. You can change the active graph by calling the \link[cgraph]{cg.active} method on another cgraph object.
}
\author{
Ron Triepels
}
