% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.linear.grad.z}
\alias{cg.linear.grad.z}
\title{Linear Transformation Gradient}
\usage{
cg.linear.grad.z(z, grad)
}
\arguments{
\item{z}{numeric vector or array, value of \code{z}.}

\item{grad}{numeric vector or array, gradient of \code{z}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x \%*\% y + as.numeric(z)} with respect to \code{z}.
}
\author{
Ron Triepels
}
\keyword{internal}
