% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.log}
\alias{cg.log}
\title{Logarithmic Function}
\usage{
cg.log(x, base, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{base}{cg.node, placeholder for a numeric scalar.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{log(x, base)}.
}
\author{
Ron Triepels
}
