% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.log2}
\alias{cg.log2}
\title{Logarithmic Base 2 Function}
\usage{
cg.log2(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{log2(x)}.
}
\author{
Ron Triepels
}
