% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{cg.logical.grad}
\alias{cg.logical.grad}
\title{Logical Gradient}
\usage{
cg.logical.grad(x)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of logical \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
