% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.mat.mul}
\alias{cg.mat.mul}
\title{Matrix Multiplication}
\usage{
cg.mat.mul(x, y, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{y}{cg.node, placeholder for a numeric matrix.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x \%*\% y}.
}
\author{
Ron Triepels
}
