% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.max}
\alias{cg.max}
\title{Maxima}
\usage{
cg.max(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{max(x)}.
}
\author{
Ron Triepels
}
