% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.mean}
\alias{cg.mean}
\title{Arithmetic Mean}
\usage{
cg.mean(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sum(x) / length(x)}.
}
\note{
For computational efficiency, this function does not use the standard \code{mean} function.
}
\author{
Ron Triepels
}
