% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.min}
\alias{cg.min}
\title{Minima}
\usage{
cg.min(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{min(x)}.
}
\author{
Ron Triepels
}
