% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.name}
\alias{cg.name}
\title{Generate Name}
\value{
character scalar, auto-generated name for the node.
}
\description{
Generate a default name for a node.
}
\details{
\code{$name()}
}
\note{
The auto-generated name is not guaranteed to be unique.

There is a wrapper function \link[cgraph]{name} that calls this method on the current active graph.
}
\examples{
# Initialize a new computational graph.
x <- cgraph$new()

# Generate a name.
x$name()

}
\author{
Ron Triepels
}
