% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.neg}
\alias{cg.neg}
\title{Negative}
\usage{
cg.neg(x, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{-x}.
}
\author{
Ron Triepels
}
