% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{cg.not}
\alias{cg.not}
\title{Not}
\usage{
cg.not(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{!x}.
}
\author{
Ron Triepels
}
