% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.opr}
\alias{cg.opr}
\title{Add Operation}
\arguments{
\item{call}{expression or call, operations performed by the node.}

\item{grads}{named list of expressions or calls, gradients of the inputs and parameters used in \code{call} with respect to the node.}

\item{binding}{named list or environment, binds the values used in the expressions or calls of \code{call} and \code{grads} to the symbols of the nodes in the graph.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, the node is tried to be added to the graph under an auto-generated name.}
}
\value{
cg.node, expression node.
}
\description{
Add an operation to the graph.
}
\details{
\code{$expr(call, grads, binding, name)}
}
\note{
The operation to be performed by the node should be provided as an expression or call to \code{call}. If this operation depends on any inputs, then the gradients of the node with respect to these inputs should be supplied as an expression or call to \code{gradients}. These gradients must be a function of each input's gradient. The special reserved word \code{grad} evaluates to this gradient at run-time and can be used in the expression and call of each input's gradient as placeholder.

Any inputs in the expressions or calls of the node (both supplied to \code{call} and \code{gradients}) should be bind to the symbols of the nodes in the graph. This can be done by supplying the names of the inputs and the corresponding nodes in the graph to which the inputs should bind to \code{binding}. At run-time, the symbols of the nodes in the graph are substituted for the inputs in the expressions or calls.

The name of the node cannot be \code{grad} as this is a reserved word.
}
\author{
Ron Triepels
}
