% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.plot}
\alias{cg.plot}
\title{Plot}
\arguments{
\item{...}{additional arguments that can be passed on to the plot function of the Rgraphiz package.}
}
\value{
none.
}
\description{
Plot the topology of the computational graph.
}
\details{
\code{$plot(...)}
}
\note{
A visual representation of the computational graph might be usefull for debugging purposes. This functions requires the Rgraphviz package.
}
\author{
Ron Triepels
}
