% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.pmax.grad.y}
\alias{cg.pmax.grad.y}
\title{Parallel Maxima Gradient}
\usage{
cg.pmax.grad.y(x, y, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{grad}{numeric vector or array, gradient of \code{y}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{pmax(x, y)} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
