% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.pmax}
\alias{cg.pmax}
\title{Parallel Maxima}
\usage{
cg.pmax(x, y, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{pmax(x, y)}.
}
\note{
In contrast to the base \code{pmax} function, this function only accepts two variables.
}
\author{
Ron Triepels
}
