% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.pmin.grad.x}
\alias{cg.pmin.grad.x}
\title{Parallel Minima Gradient}
\usage{
cg.pmin.grad.x(x, y, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{pmin(x, y)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
