% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.pmin}
\alias{cg.pmin}
\title{Parallel Minima}
\usage{
cg.pmin(x, y, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{pmin(x, y)}.
}
\note{
In contrast to the base \code{pmin} function, this function only accepts two variables.
}
\author{
Ron Triepels
}
