% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.prod.grad}
\alias{cg.prod.grad}
\title{Product of Vector Elements Gradient}
\usage{
cg.prod.grad(x, y, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{prod(x)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{prod(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
