% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.prod}
\alias{cg.prod}
\title{Product of Vector Elements}
\usage{
cg.prod(x, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{prod(x)}.
}
\note{
In contrast to the base \code{prod} function, this function only accepts a single variable.
}
\author{
Ron Triepels
}
