% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cg.reshape}
\alias{cg.reshape}
\title{Reshape Array Dimensions}
\usage{
cg.reshape(x, dims, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{dims}{numeric scalar or vector, the dimensions of the new array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Change the dimensions of array \code{x} to \code{dims}.
}
\note{
The elements of \code{x} are rearranged column-wise.
}
\author{
Ron Triepels
}
