% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.rowSums}
\alias{cg.rowSums}
\title{Row Sums}
\usage{
cg.rowSums(x, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{rowSums(x)}.
}
\author{
Ron Triepels
}
