% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.set}
\alias{cg.set}
\title{Change the Default Value of a Node}
\arguments{
\item{name}{cg.node, name of the node whose default value is to be changed.}

\item{value}{numeric vector or array, default value of the node.}
}
\value{
nothing.
}
\description{
Change the default value of a node in the graph.
}
\details{
\code{$set(name. value)}
}
\author{
Ron Triepels
}
