% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cg.sigmoid.grad}
\alias{cg.sigmoid.grad}
\title{Sigmoid Gradient}
\usage{
cg.sigmoid.grad(y, grad)
}
\arguments{
\item{y}{numeric vector or array, value of \code{1 / (1 + exp(-x))}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{1 / (1 + exp(-x))} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
