% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cg.sigmoid}
\alias{cg.sigmoid}
\title{Sigmoid}
\usage{
cg.sigmoid(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{1 / (1 + exp(1)^-x)}.
}
\note{
\code{exp(1)=e}.
}
\author{
Ron Triepels
}
