% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg.sub.grad}
\alias{cg.sub.grad}
\title{Subtract Gradient}
\usage{
cg.sub.grad(y, grad)
}
\arguments{
\item{grad}{numeric vector or array, gradient of \code{y}.}

\item{x}{numeric vector or array, value of \code{y}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x - y} with respect to \code{y}.
}
\note{
The gradient of \code{x - y} with respect to \code{x} can be calculated by \link[cgraph]{add.grad}.
}
\author{
Ron Triepels
}
\keyword{internal}
