% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.sum}
\alias{cg.sum}
\title{Sum of Vector Elements}
\usage{
cg.sum(x, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sum(x)}.
}
\note{
In contrast to the original \code{sum} function, this function does not accept a variable amount of arguments.
}
\author{
Ron Triepels
}
