% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cg.t}
\alias{cg.t}
\title{Matrix Transpose}
\usage{
cg.t(x, name)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Perform \code{t(x)}.
}
\author{
Ron Triepels
}
