% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cg.tanh.grad}
\alias{cg.tanh.grad}
\title{Hyperbolic Tangent Gradient}
\usage{
cg.tanh.grad(y, grad)
}
\arguments{
\item{y}{numeric vector or array, value of \code{tanh(x)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{tanh(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
