% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_acos}
\alias{cg_acos}
\title{Inverse Cosine}
\usage{
cg_acos(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{acos(x)}.
}
\seealso{
\link[base:Trig]{acos}
}
\author{
Ron Triepels
}
