% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg_adj_mat}
\alias{cg_adj_mat}
\title{Adjacency Matrix}
\value{
numeric matrix, the adjacency matrix of the graph.
}
\description{
Retrieve the adjacency matrix of the graph.
}
\details{
\code{$adj.mat()}
}
\author{
Ron Triepels
}
