% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_as_double}
\alias{cg_as_double}
\title{Coerce to a Numerical Vector}
\usage{
cg_as_double(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Coerce \code{x} to a one-dimensional numerical vector.
}
\note{
This function is identical to \code{cg_as_numeric}.
}
\seealso{
\link[base:double]{as.double}
}
\author{
Ron Triepels
}
