% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cg_as_numeric}
\alias{cg_as_numeric}
\title{Coerce to a Numeric Vector}
\usage{
cg_as_numeric(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Coerce \code{x} to a one-dimensional numeric vector.
}
\note{
This function is identical to \code{cg.as.double}.
}
\seealso{
\link[base:double]{as.numeric}
}
\author{
Ron Triepels
}
