% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_atan}
\alias{cg_atan}
\title{Inverse Tangent}
\usage{
cg_atan(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{atan(x)}.
}
\seealso{
\link[base:Trig]{atan}
}
\author{
Ron Triepels
}
