% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_cbrt}
\alias{cg_cbrt}
\title{Cubic Root}
\usage{
cg_cbrt(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x^(1/3)}.
}
\author{
Ron Triepels
}
