% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_colmeans}
\alias{cg_colmeans}
\title{Column Means}
\usage{
cg_colmeans(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{colMeans(x)}.
}
\note{
Function \link[base:colSums]{colMeans} is called without changing the default value of argument \code{na.rm} and \code{dims}.
}
\seealso{
\link[base:colSums]{colMeans}
}
\author{
Ron Triepels
}
