% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_colSums}
\alias{cg_colSums}
\title{Column Sums}
\usage{
cg_colSums(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{colSums(x)}.
}
\note{
Function \link[base:colSums]{colSums} is called without setting argument \code{na.rm} and \code{dims}.
}
\seealso{
\link[base:colSums]{colSums}
}
\author{
Ron Triepels
}
