% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{cg_constant}
\alias{cg_constant}
\title{Add Constant}
\usage{
cg_constant(value, name = NULL)
}
\arguments{
\item{value}{numeric vector or array, value of the node.}

\item{name}{character scalar, name of the node (optional). In case argument \code{name} is missing, the node is added to the graph under an automatically generated name.}
}
\value{
cg_node object.
}
\description{
Add a constant node to the active graph.
}
\note{
Argument \code{name} is deprecated and will be removed in the next major release.

Constant nodes are ignored when differentiating a graph.
}
\examples{
# Initialize a computational graph
graph <- cg_graph()

# Add a constant with value 1 and name 'a' to the graph.
a <- cg_constant(1, name = "a")

}
\author{
Ron Triepels
}
