% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_cos}
\alias{cg_cos}
\title{Cosine}
\usage{
cg_cos(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{cos(x)}.
}
\seealso{
\link[base:Trig]{cos}
}
\author{
Ron Triepels
}
