% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_cosh}
\alias{cg_cosh}
\title{Hyperbolic Cosine}
\usage{
cg_cosh(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{cosh(x)}.
}
\seealso{
\link[base:Hyperbolic]{cosh}
}
\author{
Ron Triepels
}
