% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_crossprod}
\alias{cg_crossprod}
\title{Matrix Crossproduct}
\usage{
cg_crossprod(x, y = NULL, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix.}

\item{y}{either a cg_node object or a numerical matrix (optional).}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{crossprod(x, y)}.
}
\seealso{
\link[base:crossprod]{crossprod}
}
\author{
Ron Triepels
}
