% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\docType{data}
\name{cg_default_library}
\alias{cg_default_library}
\title{Default Function Library}
\format{An object of class \code{environment} of length 119.}
\usage{
cg_default_library
}
\description{
This is the default function library used by a computational graph. It implements all graph operators by base R functions.
}
\keyword{datasets}
