% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_dim}
\alias{cg_dim}
\title{Dimensions of an Array}
\usage{
cg_dim(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{dim(x)}.
}
\note{
This is a non-differentiable operator.
}
\seealso{
\link[base:dim]{dim}
}
\author{
Ron Triepels
}
