% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_exp2}
\alias{cg_exp2}
\title{Base-2 Exponential Function}
\usage{
cg_exp2(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{2^x}.
}
\author{
Ron Triepels
}
