% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{cg_graph}
\alias{cg_graph}
\title{Computational Graph}
\usage{
cg_graph()
}
\value{
cg_graph object.
}
\description{
Initialize a computational graph
}
\note{
The graph is automatically set to be the active graph.
}
\examples{
# Initialize a computational graph
graph <- cg_graph()

}
\author{
Ron Triepels
}
