% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{cg_init_gaussian}
\alias{cg_init_gaussian}
\title{Add Parameter}
\usage{
cg_init_gaussian(..., mean = 0, sd = 1, name = NULL)
}
\arguments{
\item{...}{numerical scalars, the dimensions of the parameter.}

\item{mean}{numerical scalar, the mean of the Gaussian distribution (optional).}

\item{sd}{numerical scalar, the standard deviation of the Gaussian distribution (optional).}

\item{name}{character scalar, name of the node (optional). In case argument \code{name} is missing, the node is added to the graph under an automatically generated name.}
}
\value{
cg_node object.
}
\description{
Add a parameter to the active graph whose value is initialized by a Gaussian distribution.
}
\note{
The dimensions provided to \code{...} must be non-negative whole numbers.
}
\examples{
# Initialize a computational graph
graph <- cg_graph()

# Add a parameter to the graph
a <- cg_init_gaussian(2, 3, mean = 0, sd = 1, name = "a")

}
\author{
Ron Triepels
}
