% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{cg_init_xavier_uniform}
\alias{cg_init_xavier_uniform}
\title{Add Parameter}
\usage{
cg_init_xavier_uniform(..., name = NULL)
}
\arguments{
\item{...}{numerical scalars, the dimensions of the parameter.}

\item{name}{character scalar, name of the node (optional). In case argument \code{name} is missing, the node is added to the graph under an automatically generated name.}
}
\value{
cg_node object.
}
\description{
Add a parameter to the active graph whose value is initialized by Xavier initialization using a uniform distribution.
}
\note{
The dimensions provided to \code{...} must be non-negative whole numbers.
}
\examples{
# Initialize a computational graph
graph <- cg_graph()

# Add a parameter to the graph
a <- cg_init_xavier_uniform(2, 3, name = "a")

}
\author{
Ron Triepels
}
