% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_length}
\alias{cg_length}
\title{Length of an Object}
\usage{
cg_length(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or an R object.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{length(x)}.
}
\note{
This is a non-differentiable operator.
}
\seealso{
\link[base:length]{length}
}
\author{
Ron Triepels
}
