% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_less}
\alias{cg_less}
\title{Less}
\usage{
cg_less(x, y, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x < y}.
}
\seealso{
\link[base:Comparison]{less}
}
\author{
Ron Triepels
}
