% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_linear}
\alias{cg_linear}
\title{Linear Transformation}
\usage{
cg_linear(x, y, z, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix.}

\item{y}{either a cg_node object or a numerical matrix.}

\item{z}{either a cg_node object or a numerical vector.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x \%*\% y + c(z)}.
}
\note{
This function is deprecated and will be removed in the next major release. Please use function \link[cgraph:cg_linear1]{cg_linear1} instead.
}
\author{
Ron Triepels
}
