% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_linear1}
\alias{cg_linear1}
\title{Linear Transformation}
\usage{
cg_linear1(x1, y1, z = NULL, name = NULL)
}
\arguments{
\item{x1}{either a cg_node object or a numerical matrix.}

\item{y1}{either a cg_node object or a numerical matrix.}

\item{z}{either a cg_node object or a numerical vector (optional).}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x1 \%*\% y1 + c(z)}.
}
\note{
This function is equivalent to \code{cg_matmul(x1, y1) + cg_as_numeric(z)} but shorter and more efficient.
}
\author{
Ron Triepels
}
