% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_linear2}
\alias{cg_linear2}
\title{Linear Transformation}
\usage{
cg_linear2(x1, y1, x2, y2, z = NULL, name = NULL)
}
\arguments{
\item{x1}{either a cg_node object or a numerical matrix.}

\item{y1}{either a cg_node object or a numerical matrix.}

\item{x2}{either a cg_node object or a numerical matrix.}

\item{y2}{either a cg_node object or a numerical matrix.}

\item{z}{either a cg_node object or a numerical vector (optional).}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x1 \%*\% y1 + x2 \%*\% y2 + c(z)}.
}
\note{
This function is equivalent to \code{cg_matmul(x1, y1) + cg_matmul(x2, y2) + cg_as_numeric(z)} but shorter and more efficient.
}
\author{
Ron Triepels
}
