% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_ln}
\alias{cg_ln}
\title{Natural Logarithm}
\usage{
cg_ln(n1, name = NULL)
}
\arguments{
\item{name}{character scalar, name of the operation (optional).}

\item{x}{either a cg_node object or a numeric vector or array.}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{log(x)}.
}
\seealso{
\link[base:log]{log}
}
\author{
Ron Triepels
}
