% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_log10}
\alias{cg_log10}
\title{Logarithmic Base 10}
\usage{
cg_log10(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{log10(x)}.
}
\seealso{
\link[base:log]{log10}
}
\author{
Ron Triepels
}
