% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_log2}
\alias{cg_log2}
\title{Logarithm Base 2}
\usage{
cg_log2(n1, name = NULL)
}
\arguments{
\item{name}{character scalar, name of the operation (optional).}

\item{x}{either a cg_node object or a numeric vector or array.}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{log2(x)}.
}
\seealso{
\link[base:log]{log2}
}
\author{
Ron Triepels
}
