% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_matmul}
\alias{cg_matmul}
\title{Matrix Multiplication}
\usage{
cg_matmul(x, y, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix.}

\item{y}{either a cg_node object or a numerical matrix.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x \%*\% y}.
}
\seealso{
\link[base:matmult]{matmult}
}
\author{
Ron Triepels
}
