% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_matrix}
\alias{cg_matrix}
\title{Matrices}
\usage{
cg_matrix(
  data = NA,
  nrow = 1,
  ncol = 1,
  byrow = FALSE,
  dimnames = NULL,
  name = NULL
)
}
\arguments{
\item{data}{either a cg_node object or a numerical vector.}

\item{nrow}{either a cg_node object or a numerical scalar.}

\item{ncol}{either a cg_node object or a numerical scalar.}

\item{byrow}{either a cg_node object or a logical scalar.}

\item{dimnames}{either a cg_node object or a character vector.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Create a matrix.
}
\note{
This is a non-differentiable operator.
}
\seealso{
\link[base:matrix]{matrix}
}
\author{
Ron Triepels
}
