% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_max}
\alias{cg_max}
\title{Maxima}
\usage{
cg_max(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{max(x)}.
}
\note{
Function \link[base:max]{max} is called without setting argument \code{na.rm}.
}
\seealso{
\link[base:max]{max}
}
\author{
Ron Triepels
}
