% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_min}
\alias{cg_min}
\title{Minima}
\usage{
cg_min(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{min(x)}.
}
\note{
Function \link[base:Extremes]{min} is called without setting argument \code{na.rm}.
}
\seealso{
\link[base:Extremes]{min}
}
\author{
Ron Triepels
}
