% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_ncol}
\alias{cg_ncol}
\title{Number of Columns of an Array}
\usage{
cg_ncol(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{ncol(x)}.
}
\note{
This is a non-differentiable operator.
}
\seealso{
\link[base:nrow]{ncol}
}
\author{
Ron Triepels
}
