% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_neg}
\alias{cg_neg}
\title{Negative}
\usage{
cg_neg(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{-x}.
}
\seealso{
\link[base:Arithmetic]{negative}
}
\author{
Ron Triepels
}
