% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_not_equal}
\alias{cg_not_equal}
\title{Not equal}
\usage{
cg_not_equal(x, y, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x != y}.
}
\seealso{
\link[base:Comparison]{not equal}
}
\author{
Ron Triepels
}
