% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg_opr}
\alias{cg_opr}
\title{Add Operation}
\arguments{
\item{call}{symbol, operation performed by the node.}

\item{grads}{list of symbols, gradients functions of the input nodes that are consumed by the operation in argument \code{call}.}

\item{args}{list of cg_node objects, the nodes that are consumed by the operation in argument \code{call}.}

\item{name}{character scalar, name of the node (optional). In case argument \code{name} is missing, the node is tried to be added to the graph under an auto-generated name.}
}
\value{
cg.node, operation.
}
\description{
Add an operation node to the graph.
}
\details{
\code{$opr(call, grads, binding, name = NULL)}
}
\note{
The operation to be performed by the node should be provided as a symbol to argument \code{call}. If this operation consumes any other nodes in the graph, then the gradient function of the operation with respect to these input nodes should be supplied as a symbol to argument \code{gradients}. These gradients must be a function of each input's gradient. A gradient function must be provided for each input node as specified by argument \code{args}.

There is a wrapper function \link[cgraph]{opr} that calls this method on the current active graph.
}
\author{
Ron Triepels
}
