% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{cg_optim_adagrad}
\alias{cg_optim_adagrad}
\title{Adaptive Gradient (AdaGrad)}
\usage{
cg_optim_adagrad(parms = list(), eta = 0.01, eps = 1e-08)
}
\arguments{
\item{parms}{list of cg_node objects, the nodes to be optimzed.}

\item{eta}{numeric scalar, learning rate.}

\item{eps}{numeric scalar, small term to improve numerical stability (optional).}
}
\value{
cg_optim object.
}
\description{
Initialize an AdaGrad optimizer.
}
\author{
Ron Triepels
}
