% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{cg_optim_gd}
\alias{cg_optim_gd}
\title{Gradient Descent}
\usage{
cg_optim_gd(parms = list(), eta = 0.05)
}
\arguments{
\item{parms}{list of cg_node objects, the nodes to be optimzed.}

\item{eta}{numeric scalar, learning rate.}
}
\value{
cg_optim object.
}
\description{
Initialize a gradient descent optimizer.
}
\author{
Ron Triepels
}
