% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{cg_optim_step}
\alias{cg_optim_step}
\title{Optimization Step}
\usage{
cg_optim_step(optim)
}
\arguments{
\item{optim}{cg_optim object, the optimizer to be used for the optimization.}
}
\value{
cg_optim object.
}
\description{
Perform a single optimization step.
}
\author{
Ron Triepels
}
