% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{cg_parameter}
\alias{cg_parameter}
\title{Add Parameter}
\usage{
cg_parameter(value, name = NULL)
}
\arguments{
\item{value}{numeric vector or array, value of the node.}

\item{name}{character scalar, name of the node (optional). In case argument \code{name} is missing, the node is added to the graph under an automatically generated name.}
}
\value{
cg_node object.
}
\description{
Add a parameter node to the active graph.
}
\note{
Parameters are assumed to be subject to some optimization process. Hence, their value might change over time.
}
\examples{
# Initialize a computational graph
x <- cg_graph()

# Add a parameter with value 1 and name 'a' to the graph.
a <- cg_parameter(1, name = "a")

}
\author{
Ron Triepels
}
