% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_pmax}
\alias{cg_pmax}
\title{Parallel Maxima}
\usage{
cg_pmax(x, y, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{pmax(x, y)}.
}
\note{
Function \link[base:pmax]{pmax} is called without setting argument \code{na.rm}.
}
\seealso{
\link[base:pmax]{pmax}
}
\author{
Ron Triepels
}
