% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_pmin}
\alias{cg_pmin}
\title{Parallel Minima}
\usage{
cg_pmin(x, y, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{pmin(x, y)}.
}
\note{
Function \link[base:Extremes]{pmin} is called without setting argument \code{na.rm}.
}
\seealso{
\link[base:Extremes]{pmin}
}
\author{
Ron Triepels
}
