% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_pos}
\alias{cg_pos}
\title{Positive}
\usage{
cg_pos(n1, name = NULL)
}
\arguments{
\item{name}{character scalar, name of the operation (optional).}

\item{x}{either a cg_node object or a numeric vector or array.}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{x}.
}
\seealso{
\link[base:Arithmetic]{positive}
}
\author{
Ron Triepels
}
