% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_reshape}
\alias{cg_reshape}
\title{Reshape Array Dimensions}
\usage{
cg_reshape(x, dim, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{dim}{cg.node, placeholder for a numeric scalar or vector, the dimensions of the new array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Change the dimensions of array \code{x} to \code{dim}.
}
\note{
The elements of \code{x} are re-arranged column-wise by base function \link[base:array]{array}.
}
\seealso{
\link[base:array]{array}
}
\author{
Ron Triepels
}
