% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_rowSums}
\alias{cg_rowSums}
\title{Row Sums}
\usage{
cg_rowSums(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{rowSums(x)}.
}
\note{
Function \link[base:colSums]{rowSums} is called without setting argument \code{na.rm} and \code{dims}.
}
\seealso{
\link[base:colSums]{rowSums}
}
\author{
Ron Triepels
}
