% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{cg_session_graph}
\alias{cg_session_graph}
\title{Get Active Graph}
\usage{
cg_session_graph()
}
\value{
cg_graph object, the graph that is currently active.
}
\description{
Get the graph that is currently active.
}
\examples{
# Initialize a computational graph
x <- cg_graph()

# Retrieve the graph from the session
cg_session_graph()

}
\author{
Ron Triepels
}
