% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cg_sigmoid}
\alias{cg_sigmoid}
\title{Sigmoid}
\usage{
cg_sigmoid(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sigmoid(x)}.
}
\author{
Ron Triepels
}
