% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_sinh}
\alias{cg_sinh}
\title{Hyperbolic Sine}
\usage{
cg_sinh(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{sinh(x)}.
}
\seealso{
\link[base:Hyperbolic]{sinh}
}
\author{
Ron Triepels
}
