% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_sqrt}
\alias{cg_sqrt}
\title{Square Root}
\usage{
cg_sqrt(n1, name = NULL)
}
\arguments{
\item{name}{character scalar, name of the operation (optional).}

\item{x}{either a cg_node object or a numeric vector or array.}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{sqrt(x)}.
}
\seealso{
\link[base:MathFun]{sqrt}
}
\author{
Ron Triepels
}
