% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_square}
\alias{cg_square}
\title{Square}
\usage{
cg_square(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x ^ 2}.
}
\note{
This function is equivalent to \code{cg_pow(x, 2)}.
}
\seealso{
\link[base:Arithmetic]{square}
}
\author{
Ron Triepels
}
