% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{cg_sub_assign1}
\alias{cg_sub_assign1}
\title{Subassign}
\usage{
cg_sub_assign1(x, ..., y, name = NULL)
}
\arguments{
\item{x}{a cg_node object.}

\item{...}{either cg_node objects or numerical scalars that are passed on to the \code{`[<-`} function.}

\item{y}{either a cg_node or an array of a similar class as \code{x}.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x[...] <- y}.
}
\note{
This operator is not differentiable with respect to the arguments provided to \code{...}. Any attempt to differentiate this operator with respect to these arguments results in an error.
}
\seealso{
\link[base:Extract]{subassign}
}
\author{
Ron Triepels
}
