% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{cg_sub_assign2}
\alias{cg_sub_assign2}
\title{Subassign}
\usage{
cg_sub_assign2(x, i, y, name = NULL)
}
\arguments{
\item{x}{a cg_node object.}

\item{i}{either a cg_node or a numerical scalar that is passed on to the \code{`[[<-`} function.}

\item{y}{either a cg_node or an array of a similar class as \code{x}.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x[[i]] <- y}.
}
\note{
This operator is not differentiable with respect to \code{i}. Any attempt to differentiate this operator with respect to this argument results in an error.
}
\seealso{
\link[base:Extract]{subassign}
}
\author{
Ron Triepels
}
