% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{cg_subset2}
\alias{cg_subset2}
\title{Subset}
\usage{
cg_subset2(x, ..., name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{...}{either cg_node objects or numerical scalars that are passed on to the \code{`[[`} function.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{x[[...]]}.
}
\note{
This operator is not differentiable with respect to the arguments provided to \code{...}. Any attempt to differentiate this operator with respect to these arguments results in an error.
}
\seealso{
\link[base:Extract]{subset}
}
\author{
Ron Triepels
}
