% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_sum}
\alias{cg_sum}
\title{Sum of Vector Elements}
\usage{
cg_sum(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{sum(x)}.
}
\note{
In contrast to the base \link[base:sum]{sum} function, this function only accepts a single argument.

Function \link[base:sum]{sum} is called without changing the default value of argument \code{na.rm}.
}
\seealso{
\link[base:sum]{sum}
}
\author{
Ron Triepels
}
