% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_t}
\alias{cg_t}
\title{Matrix Transpose}
\usage{
cg_t(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Perform \code{t(x)}.
}
\seealso{
\link[base:t]{t}
}
\author{
Ron Triepels
}
