% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cg_tan}
\alias{cg_tan}
\title{Tangent}
\usage{
cg_tan(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{tan(x)}.
}
\seealso{
\link[base:Trig]{tan}
}
\author{
Ron Triepels
}
