% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cg_tanh}
\alias{cg_tanh}
\title{Hyperbolic Tangent}
\usage{
cg_tanh(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{tanh(x)}.
}
\seealso{
\link[base:Hyperbolic]{tanh}
}
\author{
Ron Triepels
}
