% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_tcrossprod}
\alias{cg_tcrossprod}
\title{Transpose Matrix Crossproduct}
\usage{
cg_tcrossprod(x, y = NULL, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{y}{cg.node, placeholder for a numeric matrix (optional).}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{tcrossprod(x, y)}.
}
\seealso{
\link[base:crossprod]{tcrossprod}
}
\author{
Ron Triepels
}
