% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_vector}
\alias{cg_vector}
\title{Vectors}
\usage{
cg_vector(mode = "logical", length = 0L, name = NULL)
}
\arguments{
\item{mode}{either a cg_node object or a character scalar.}

\item{length}{either a cg_node object or a numeric vector.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Create a vector.
}
\note{
This operator is not differentiable. Any attempt to differentiate this operator will result in an error.
}
\seealso{
\link[base:vector]{vector}
}
\author{
Ron Triepels
}
