% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cgraph.adj.mat}
\alias{cgraph.adj.mat}
\title{Adjacency Matrix}
\value{
numeric matrix, the adjacency matrix of the graph.
}
\description{
Create an adjacency matrix of the computational graph.
}
\details{
\code{$adj.mat()}
}
\author{
Ron Triepels
}
