% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cgraph}
\alias{cgraph}
\title{Computational Graph Class Generator}
\description{
Generator class to create new cgraph objects.
}
\section{Usage}{

\preformatted{cgraph}
}

\section{Methods}{

\describe{
\item{\code{$new()}}{initialize a new computational graph.}
}
}

\author{
Ron Triepels
}
