% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{colMeans}
\alias{colMeans}
\title{Column Means}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{colMeans(x)}.
}
\section{Usage}{

\preformatted{colMeans(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric array of two or more dimensions.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
