% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{colSums}
\alias{colSums}
\title{Column Sums}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{colSums(x)}.
}
\section{Usage}{

\preformatted{colSums(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
