% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cos.cg.node}
\alias{cos.cg.node}
\title{Cosinus}
\usage{
\method{cos}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{cos(x)}.
}
\author{
Ron Triepels
}
