% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cos}
\alias{cos}
\title{Cosinus}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{cos(x)}.
}
\section{Usage}{

\preformatted{cos(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
