% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cos_grad}
\alias{cos_grad}
\title{Cosine Gradient}
\usage{
cos_grad(x, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{val}{numeric vector or array, value of \code{cos(x)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{cos(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
