% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{cosh.cg.node}
\alias{cosh.cg.node}
\title{Hyperbolic Cosine}
\usage{
\method{cosh}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{cosh(x)}.
}
\author{
Ron Triepels
}
