% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{count.type}
\alias{count.type}
\title{Count Nodes}
\value{
numeric scalar, the numer of nodes of the given type.
}
\description{
Count how many nodes of a specific type are added to the graph.
}
\section{Usage}{

\preformatted{count.type(type)}
}

\section{Agruments}{

\describe{
\item{type}{numeric scalar, type of the node. Should be either: 0 (constant), 1 (input), 2 (parameter), or 3 (expression). Defaults to 3 (expression).}
}
}

\author{
Ron Triepels
}
