% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{crossprod_grad_y}
\alias{crossprod_grad_y}
\title{Matrix Crossproduct Gradient}
\usage{
crossprod_grad_y(x, y, val, grad)
}
\arguments{
\item{x}{numeric matrix, value of \code{x}.}

\item{y}{numeric matrix, value of \code{y}.}

\item{val}{numeric matrix, value of \code{crossprod(x, y)}.}

\item{grad}{numeric matrix, gradient of \code{y}.}
}
\value{
numeric matrix, gradient of the operation.
}
\description{
Calculate the gradient of \code{crossprod(x, y)} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
