% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{d}
\alias{d}
\title{Dimensions of an Vector or Array}
\usage{
d(x)
}
\arguments{
\item{x}{vector or array, the object whose dimensions need to be retrieved.}
}
\value{
numeric scalar, the dimensions of the vector or array.
}
\description{
Retrieve the dimension of an Vector or Array
}
\note{
If \code{x} is an array, the function will return \code{dim(x)}. Otherwise, the function returns \code{length(x)}.
}
\author{
Ron Triepels
}
