% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{div.grad.y}
\alias{div.grad.y}
\title{Divide Gradient}
\usage{
div.grad.y(x, y, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{grad}{numeric vector or array, gradient of \code{y}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x / y} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
