% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{divide}
\alias{divide}
\title{Divide}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{x / y}.
}
\section{Usage}{

\preformatted{divide(x, y, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{y}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
