% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{equal_grad_x}
\alias{equal_grad_x}
\title{Equal Gradient}
\usage{
equal_grad_x(x, y, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{val}{numeric vector or array, value of \code{x == y}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x == y} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
