% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{exp.cg.node}
\alias{exp.cg.node}
\title{Exponential Function}
\usage{
\method{exp}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{exp(1)^x}.
}
\note{
\code{exp(1)=e}.
}
\author{
Ron Triepels
}
