% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{exp}
\alias{exp}
\title{Exponential Function}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{exp(1)^x}.
}
\note{
\code{exp(1)=e}.
}
\section{Usage}{

\preformatted{exp(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
