% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{export}
\alias{export}
\title{Export Function}
\usage{
export(name, fun)
}
\arguments{
\item{name}{character scalar, the name to which the function is to be binded in the \code{functions} environment.}

\item{fun}{function, the function to be exported.}
}
\value{
nothing
}
\description{
Export a function to the \code{functions} environment in the global session.
}
\author{
Ron Triepels
}
\keyword{internal}
