% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{expr}
\alias{expr}
\title{Add Expression}
\usage{
expr(call, grads, binding, name)
}
\arguments{
\item{call}{expression or call, operations performed by the node.}

\item{grads}{named list of expressions or calls, gradients of the inputs and parameters used in \code{call} with respect to the node.}

\item{binding}{named list or environment, binds the values used in the expressions or calls of \code{call} and \code{grads} to the symbols of the nodes in the graph.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, an auto-generated name is assigned to the node.}
}
\value{
cg.node, expression node.
}
\description{
Add an expression node to the active graph.
}
\note{
The operation that is to be performed by the node should be provided as an expression or call to argument \code{call}. If this operation contains any inputs or parameters, then the gradient of these inputs and parameters with respect to the node should be provided via argument \code{gradients}. Also, any variables used in the these expressions or calls should be bind to the symbols of the nodes in the graph. There are two ways to bind variables. Either, \code{binding} should be a named list were the names of nodes are assigned as symbols to the named members, or \code{binding} is an environment were the names of nodes are assigned as symbols to objects within the environment.
}
\author{
Ron Triepels
}
