% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{\%:\%}
\alias{\%:\%}
\title{Assign a Value to a Node}
\usage{
x \%:\% value
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}

\item{value}{numeric scalar or array, default value of the node.}
}
\value{
nothing.
}
\description{
Assign a default value to a node in the current graph.
}
