% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{\%mul\%}
\alias{\%mul\%}
\title{Matrix Multiplication}
\usage{
x \%mul\% y
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{y}{cg.node, placeholder for a numeric matrix.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x \%*\% y}.
}
\author{
Ron Triepels
}
