% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{initialize}
\alias{initialize}
\title{Computational Graph}
\value{
cgraph object.
}
\description{
Initialize a new computational graph.
}
\section{Usage}{

\preformatted{cgraph$new()}
}

\author{
Ron Triepels
}
