% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{input}
\alias{input}
\title{Add Input}
\value{
symbol, the name of the node.
}
\description{
Add an input node to the graph.
}
\section{Usage}{

\preformatted{input(value, name)}
}

\section{Agruments}{

\describe{
\item{value}{numeric scalar or array, default value of the node.}
\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, an auto-generated name is assigned to the node.}
}
}

\author{
Ron Triepels
}
