% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{linear_grad_y}
\alias{linear_grad_y}
\title{Linear Transformation Gradient}
\usage{
linear_grad_y(x, y, z, val, grad)
}
\arguments{
\item{x}{numeric matrix, value of \code{x}.}

\item{y}{numeric matrix, value of \code{y}.}

\item{val}{numeric matrix, value of \code{linear(x, y, z)}.}

\item{grad}{numeric matrix, gradient of \code{y}.}
}
\value{
numeric matrix, gradient of the operation.
}
\description{
Calculate the gradient of \code{linear(x, y, z)} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
