% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{linear_grad_z}
\alias{linear_grad_z}
\title{Linear Transformation Gradient}
\arguments{
\item{x}{numeric matrix, value of \code{x}.}

\item{y}{numeric matrix, value of \code{y}.}

\item{val}{numeric matrix, value of \code{linear(x, y, z)}.}

\item{grad}{numeric matrix, gradient of \code{z}.}
}
\value{
numeric matrix, gradient of the operation.
}
\description{
Calculate the gradient of \code{linear(x, y, z)} with respect to \code{z}.
}
\author{
Ron Triepels
}
\keyword{internal}
