% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{ln.grad}
\alias{ln.grad}
\title{Natural Logarithmic Gradient Gradient}
\usage{
ln.grad(x, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{log(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
