% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{ln}
\alias{ln}
\title{Natural Log}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{log(x, base = exp(1))}.
}
\note{
\code{exp(1)=e}.
}
\section{Usage}{

\preformatted{ln(x, y, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
