% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{log.cg.node}
\alias{log.cg.node}
\title{Logarithmic Function}
\usage{
\method{log}{cg.node}(x, base)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{base}{cg.node, placeholder for a numeric scalar.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{log(x, base)}.
}
\author{
Ron Triepels
}
