% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{log10.cg.node}
\alias{log10.cg.node}
\title{Logarithmic Base 10 Function}
\usage{
log10.cg.node(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{log10(x)}.
}
\author{
Ron Triepels
}
