% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{log2.cg.node}
\alias{log2.cg.node}
\title{Logarithmic Base 2 Function}
\usage{
log2.cg.node(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{log2(x)}.
}
\author{
Ron Triepels
}
