% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{log2.grad}
\alias{log2.grad}
\title{Logarithmic Base 2 Gradient Function}
\usage{
log2.grad(x, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{log2(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
