% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{mat.mul}
\alias{mat.mul}
\title{Matrix Multiplication}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{x \%*\% y}.
}
\section{Usage}{

\preformatted{mat.mul(x, y, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric matrix.}
\item{y}{character scalar or symbol, placeholder for a numeric matrix.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
