% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{matmul_grad_x}
\alias{matmul_grad_x}
\title{Matrix Multiplication Gradient}
\arguments{
\item{x}{numeric matrix, value of \code{x}.}

\item{y}{numeric matrix, value of \code{y}.}

\item{val}{numeric matrix, value of \code{x \%*\% y}.}

\item{grad}{numeric matrix, gradient of \code{x}.}
}
\value{
numeric matrix, gradient of the operation.
}
\description{
Calculate the gradient of \code{x \%*\% y} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
