% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{matmul_grad_y}
\alias{matmul_grad_y}
\title{Matrix Multiplication Gradient}
\usage{
matmul_grad_y(x, y, val, grad)
}
\arguments{
\item{x}{numeric matrix, value of \code{x}.}

\item{y}{numeric matrix, value of \code{y}.}

\item{val}{numeric matrix, value of \code{x \%*\% y}.}

\item{grad}{numeric matrix, gradient of \code{y}.}
}
\value{
numeric matrix, gradient of the operation.
}
\description{
Calculate the gradient of \code{x \%*\% y} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
