% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{mean.cg.node}
\alias{mean.cg.node}
\title{Arithmetic Mean}
\usage{
\method{mean}{cg.node}(x, name = cgraph::name(), ...)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sum(x) / length(x)}.
}
\note{
For computational efficiency, this function does not use the standard \code{mean} function.
}
\author{
Ron Triepels
}
