% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{mean.grad}
\alias{mean.grad}
\title{Arithmetic Mean Gradient}
\usage{
\method{mean}{grad}(x, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{sum(x) / length(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
