% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{mean}
\alias{mean}
\title{Arithmetic Mean}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{sum(x) / length(x)}.
}
\note{
For computational efficiency, this function does not use the standard \code{mean} function.
}
\section{Usage}{

\preformatted{mean(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
