% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{min.grad}
\alias{min.grad}
\title{Minima Gradient}
\usage{
\method{min}{grad}(x, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}

\item{y}{numeric vector or array, value of \code{min(y)}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{min(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
