% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mul.grad.y}
\alias{mul.grad.y}
\title{Multiply Gradient}
\usage{
mul.grad.y(x, y, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{grad}{numeric vector or array, gradient of \code{y}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x * y} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
