% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{name}
\alias{name}
\title{Generate Name}
\usage{
name()
}
\value{
symbol, auto-generated name for the node.
}
\description{
Generate a default name for an expression.
}
\note{
The auto-generated name is not guaranteed to be unique. Instead, it may alread be used by an expression node in the active graph.
}
\author{
Ron Triepels
}
