% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{not.grad}
\alias{not.grad}
\title{Logical Gradient}
\usage{
not.grad(x, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of logical \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
