% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{not_grad}
\alias{not_grad}
\title{Not Gradient}
\usage{
not_grad(x, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{val}{numeric vector or array, value of \code{!x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{!x}.
}
\author{
Ron Triepels
}
\keyword{internal}
