% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{perm}
\alias{perm}
\title{Array Transposition}
\usage{
perm(a)
}
\arguments{
\item{a}{array, the array to be transposed.}
}
\value{
array, transposed array.
}
\description{
Transpose the row and columns of an array while keeping its dimensions the same.
}
\note{
This is a wrapper function for \code{aperm(a, replace(1:length(dim(a)), 1:2, 2:1), resize = F)}.
}
\author{
Ron Triepels
}
