% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{plot.cgraph}
\alias{plot.cgraph}
\title{Plot}
\usage{
\method{plot}{cgraph}(x, ...)
}
\arguments{
\item{x}{cgraph object, computational graph that is to be plotted.}

\item{...}{additional arguments that can be passed on to the plot function of the Rgraphiz package.}
}
\value{
none.
}
\description{
Plot the topology of a computational graph.
}
\note{
A visual representation of the computational graph might be usefull for debugging purposes. This functions requires the Rgraphviz package.
}
\author{
Ron Triepels
}
