% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{plot}
\alias{plot}
\title{Plot}
\value{
none.
}
\description{
Plot the topology of the computational graph.
}
\note{
A visual representation of the computational graph might be usefull for debugging purposes. This functions requires the Rgraphviz package.
}
\section{Usage}{

\preformatted{plot(...)}
}

\section{Agruments}{

\describe{
\item{...}{additional arguments that can be passed on to the plot function of the Rgraphiz package.}
}
}

\author{
Ron Triepels
}
