% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{pmax_grad_x}
\alias{pmax_grad_x}
\title{Parallel Maxima Gradient}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{val}{numeric vector or array, value of \code{pmax(x, y)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{pmax(x, y)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
