% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{pmin_grad_y}
\alias{pmin_grad_y}
\title{Parallel Minima Gradient}
\usage{
pmin_grad_y(x, y, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{val}{numeric vector or array, value of \code{pmin(x, y)}.}

\item{grad}{numeric vector or array, gradient of \code{y}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{pmin(x, y)} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
