% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{reshape.grad.x}
\alias{reshape.grad.x}
\title{Reshape Array Dimensions Gradient}
\usage{
reshape.grad.x(x, dim, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{array(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
