% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{reshape}
\alias{reshape}
\title{Reshape Array Dimensions}
\value{
symbol, name of the operation.
}
\description{
Change the dimensions of array \code{x} to \code{dims}.
}
\note{
The elements of \code{x} are rearranged column-wise.
}
\section{Usage}{

\preformatted{reshape(x, dims, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric matrix.}
\item{dims}{numeric scalar or vector, the dimensions of the new array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
