% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{reshape_grad_dim}
\alias{reshape_grad_dim}
\title{Reshape Array Dimensions Gradient}
\usage{
reshape_grad_dim(x, dim, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{dim}{numeric scalar or vector, value of \code{dim}.}

\item{val}{numeric vector or array, value of \code{array(x)}.}

\item{grad}{numeric vector or array, gradient of \code{dim}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{array(x)} with respect to \code{dim}.
}
\author{
Ron Triepels
}
\keyword{internal}
