% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{rowSums_grad}
\alias{rowSums_grad}
\title{Row Sums Gradient}
\usage{
rowSums_grad(x, val, grad)
}
\arguments{
\item{x}{numeric array, value of \code{x}.}

\item{val}{numeric array, value of \code{rowSums(x)}.}

\item{grad}{numeric array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{rowSums(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
