% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid}
\arguments{
\item{x}{numeric vector or array, object on which the sigmoid function is evaluated.}
}
\value{
numeric vector or array, sigmoid values.
}
\description{
Evaluate the sigmoid function on \code{x} element-wise.
}
\note{
The sigmoid function is defined as \code{1 / (1 + exp(-x))}. The resulting sigmoid values are clamped within the range \code{[a, 1 - a]} where \code{a} evaluates to \code{.Machine$double.eps} to avoid numerical underflow.
}
\author{
Ron Triepels
}
\keyword{internal}
