% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{sin.cg.node}
\alias{sin.cg.node}
\title{Sinus}
\usage{
\method{sin}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sin(x)}.
}
\author{
Ron Triepels
}
