% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{sin_grad}
\alias{sin_grad}
\title{Sine Gradient}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{val}{numeric vector or array, value of \code{sin(x)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{sin(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
