% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{sinh.cg.node}
\alias{sinh.cg.node}
\title{Hyperbolic Sine}
\usage{
\method{sinh}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sinh(x)}.
}
\author{
Ron Triepels
}
