% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sqrt.cg.node}
\alias{sqrt.cg.node}
\title{Square Root}
\usage{
\method{sqrt}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric scalar or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sqrt(x)}.
}
\author{
Ron Triepels
}
