% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sqrt}
\alias{sqrt}
\title{Square Root}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{sqrt(x)}.
}
\section{Usage}{

\preformatted{sqrt(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
