% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sub.grad}
\alias{sub.grad}
\title{Subtract Gradient}
\usage{
sub.grad(x, y, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{y}.}

\item{grad}{numeric vector or array, gradient of \code{y}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x - y} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
