% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{sub_grad_y}
\alias{sub_grad_y}
\title{Subtract Gradient}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{y}{numeric vector or array, value of \code{y}.}

\item{val}{numeric vector or array, value of \code{x - y}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{x - y} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
