% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{t.cg.node}
\alias{t.cg.node}
\title{Matrix Transpose}
\usage{
\method{t}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}
}
\value{
cg.node, node of the operation.
}
\description{
Perform \code{t(x)}.
}
\author{
Ron Triepels
}
