% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{t}
\alias{t}
\title{Matrix Transpose}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{t(x)}.
}
\section{Usage}{

\preformatted{t(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric matrix.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
