% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{t_grad}
\alias{t_grad}
\title{Matrix Transpose Gradient}
\usage{
t_grad(x, val, grad)
}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{val}{numeric vector or array, value of \code{t(x)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{t(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
