% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{tan.cg.node}
\alias{tan.cg.node}
\title{Tangent}
\usage{
\method{tan}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{tan(x)}.
}
\author{
Ron Triepels
}
