% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{tanh.cg.node}
\alias{tanh.cg.node}
\title{Hyperbolic Tangent}
\usage{
\method{tanh}{cg.node}(x)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{tanh(x)}.
}
\author{
Ron Triepels
}
