% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{tanh.grad}
\alias{tanh.grad}
\title{Hyperbolic Tangent Gradient}
\usage{
\method{tanh}{grad}(x, val, grad)
}
\arguments{
\item{grad}{numeric vector or array, gradient of \code{x}.}

\item{y}{numeric vector or array, value of \code{tanh(x)}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{tanh(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
