% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{tanh}
\alias{tanh}
\title{Hyperbolic Tangent}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{tanh(x)}.
}
\section{Usage}{

\preformatted{tanh(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
