% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{tcrossprod}
\alias{tcrossprod}
\title{Transpose Matrix Crossproduct}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{tcrossprod(x, y)}.
}
\note{
In contrast to the original \code{tcrossprod} function, this function requires both \code{x} and \code{y} to be supplied.
}
\section{Usage}{

\preformatted{tcrossprod(x, y, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric matrix.}
\item{y}{character scalar or symbol, placeholder for a numeric matrix.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
