% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{tcrossprod_grad_x}
\alias{tcrossprod_grad_x}
\title{Transpose Matrix Crossproduct Gradient}
\usage{
tcrossprod_grad_x(x, y, val, grad)
}
\arguments{
\item{x}{numeric matrix, value of \code{x}.}

\item{y}{numeric matrix, value of \code{y}.}

\item{val}{numeric matrix, value of \code{tcrossprod(x, y)}.}

\item{grad}{numeric matrix, gradient of \code{x}.}
}
\value{
numeric matrix, gradient of the operation.
}
\description{
Calculate the gradient of \code{tcrossprod(x, y)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
