% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{tcrossprod_grad_y}
\alias{tcrossprod_grad_y}
\title{Transpose Matrix Crossproduct Gradient}
\arguments{
\item{x}{numeric matrix, value of \code{x}.}

\item{y}{numeric matrix, value of \code{y}.}

\item{val}{numeric matrix, value of \code{tcrossprod(x, y)}.}

\item{grad}{numeric matrix, gradient of \code{y}.}
}
\value{
numeric matrix, gradient of the operation.
}
\description{
Calculate the gradient of \code{tcrossprod(x, y)} with respect to \code{y}.
}
\author{
Ron Triepels
}
\keyword{internal}
