% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as.double.cg.node}
\alias{as.double.cg.node}
\title{Coerce to a Numeric Vector}
\usage{
\method{as.double}{cg.node}(x, name, ...)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
cg.node, node of the operation.
}
\description{
Coerce \code{x} to a one-dimensional numeric vector.
}
\author{
Ron Triepels
}
