% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.active}
\alias{cg.active}
\title{Change Active Graph}
\value{
none.
}
\description{
Set the graph to be the active graph.
}
\details{
\code{$active()}
}
\note{
Any placeholders or expressions that are invoked are added to the active graph. This behavior also applies to expressions that are invoked by overloaded S3 functions that do not follow the cg.<name> naming convention (such as primitive functions '+' and '-'). Moreover, only one graph can be active at a time. You can use this function to change the active graph.
}
\author{
Ron Triepels
}
