% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.approx.grad}
\alias{cg.approx.grad}
\title{Approximate Gradients}
\arguments{
\item{x}{character scalar or symbol, name of the node.}

\item{y}{character scalar or symbol, name of the node.}

\item{values}{named list or environment, values that are subsituted for the expressions and placeholders in the graph.}

\item{index}{numeric scalar, index of the target node that needs to be differentiated. Defaults to the first element.}

\item{eps}{numeric scalar, step size. Defaults to 1e-4.}
}
\value{
numeric vector or array, the derivative of \code{x} with respect to \code{y}.
}
\description{
Differentiate node \code{x} with respect to node \code{y} in the graph by numerical differentiation.
}
\details{
\code{$approx.grad(x, y, values = list(), index = 1, eps = 1e-4)}
}
\note{
All placeholders required to compute node \code{name} must have a value. Placeholders can be assigned a default value when they are created. Alternatively, argument \code{values} can be used to substitute values for placeholders that do not have a default value or to fix the values of nodes.

The graph is differentiation by the symmetric difference quotient. This function is mainly used for testing purposes.
}
\author{
Ron Triepels
}
