% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.const}
\alias{cg.const}
\title{Add Constant}
\arguments{
\item{value}{numeric scalar or array, default value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, the node is tried to be added to the graph under an auto-generated name.}
}
\value{
cg.node, constant node.
}
\description{
Add a constant node to the graph.
}
\details{
\code{$const(value, name)}
}
\note{
Constant nodes are ignored when differentiating a graph.
}
\author{
Ron Triepels
}
