% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.get.parms}
\alias{cg.get.parms}
\title{Get Parameters}
\value{
named list, parameters of the graph.
}
\description{
List all parameters and their values.
}
\details{
\code{$get.parms()}
}
\examples{
# Initialize a new computational graph.
x <- cgraph$new()

# Add some parameters.
x$add.parms(prm1 = 1, prm2 = 2, prm3 = 3)

# List the parameters.
x$get.parms()

}
\author{
Ron Triepels
}
