% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.initialize}
\alias{cg.initialize}
\title{Computational Graph}
\value{
cgraph object.
}
\description{
Initialize a new computational graph.
}
\details{
\code{$new()}
}
\note{
The cgraph object is set as the current graph. Any placeholders or expressions that are invoked will be added to this graph. You can change the current graph by calling the \code{$active()} method on another cgraph object.
}
\author{
Ron Triepels
}
