% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.initialize}
\alias{cg.initialize}
\title{Computational Graph}
\value{
cgraph object.
}
\description{
Initialize a computational graph.
}
\details{
\code{$new()}
}
\note{
The cgraph object is set to be the active graph. Any nodes that are created by wrapper function \code{const}, \code{input}, or \code{parm} will be added to this graph unless the active graph is changed. You can change the current graph by calling the \link[cgraph]{cg.active} method on another cgraph object.
}
\author{
Ron Triepels
}
