% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg.linear}
\alias{cg.linear}
\title{Linear Transformation}
\usage{
cg.linear(x, y, z, name = cgraph::name())
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{y}{cg.node, placeholder for a numeric matrix.}

\item{z}{cg.node, placeholder for a numeric vector.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x \%*\% y + c(z)}.
}
\note{
This function is equivalent to \code{add(mat.mul(x, y), c(z))}.
}
\author{
Ron Triepels
}
