% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.name}
\alias{cg.name}
\title{Generate Name}
\arguments{
\item{type}{numeric scalar, type of the node. Should be either: 0 (constant), 1 (input), 2 (parameter), or 3 (expression). Defaults to 3 (expression).}
}
\value{
symbol, auto-generated name for the node.
}
\description{
Generate a default name for a node.
}
\details{
\code{$name(type = 3)}
}
\author{
Ron Triepels
}
