% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg.parm}
\alias{cg.parm}
\title{Add Parameter}
\arguments{
\item{value}{numeric scalar or array, value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, the node is tried to be added to the graph under an auto-generated name.}
}
\value{
cg.node, parameter.
}
\description{
Add a parameter node to the graph.
}
\details{
\code{$parm(value, name)}
}
\note{
Parameters are assumed to be subject to some optimization process. Their value might change over time.

The name of the parameter node cannot be 'grad' as this is a reserved word.
}
\author{
Ron Triepels
}
