% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_abs}
\alias{cg_abs}
\title{Absolute Value}
\usage{
cg_abs(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{abs(x)}.
}
\seealso{
\link[base:MathFun]{abs}
}
\author{
Ron Triepels
}
