% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_acosh}
\alias{cg_acosh}
\title{Inverse Hyperbolic Cosine}
\usage{
cg_acosh(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{acosh(x)}.
}
\seealso{
\link[base:Hyperbolic]{acosh}
}
\author{
Ron Triepels
}
