% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_add}
\alias{cg_add}
\title{Add}
\usage{
cg_add(n1, n2, name = NULL)
}
\arguments{
\item{name}{character scalar, name of the operation (optional).}

\item{x}{either a cg_node object or a numeric vector or array.}

\item{y}{either a cg_node object or a numeric vector or array.}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{x + y}.
}
\seealso{
\link[base:Arithmetic]{add}
}
\author{
Ron Triepels
}
