% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_asin}
\alias{cg_asin}
\title{Inverse Sine}
\usage{
cg_asin(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{asin(x)}.
}
\seealso{
\link[base:Trig]{asin}
}
\author{
Ron Triepels
}
