% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_asinh}
\alias{cg_asinh}
\title{Inverse Hyperbolic Sine}
\usage{
cg_asinh(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{asinh(x)}.
}
\seealso{
\link[base:Hyperbolic]{asinh}
}
\author{
Ron Triepels
}
