% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_atanh}
\alias{cg_atanh}
\title{Inverse Hyperbolic Tangent}
\usage{
cg_atanh(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{atanh(x)}.
}
\seealso{
\link[base:Hyperbolic]{atanh}
}
\author{
Ron Triepels
}
