% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_colMeans}
\alias{cg_colMeans}
\title{Column Means}
\usage{
cg_colMeans(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{colMeans(x)}.
}
\note{
Function \link[base:colSums]{colMeans} is called without setting argument \code{na.rm} and \code{dims}.
}
\seealso{
\link[base:colSums]{colMeans}
}
\author{
Ron Triepels
}
