% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_colsums}
\alias{cg_colsums}
\title{Column Sums}
\usage{
cg_colsums(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{colSums(x)}.
}
\note{
Function \link[base:colSums]{colSums} is called without changing the default value of argument \code{na.rm} and \code{dims}.
}
\seealso{
\link[base:colSums]{colSums}
}
\author{
Ron Triepels
}
