% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_copy}
\alias{cg_copy}
\title{Copy}
\usage{
cg_copy(x, y, offset = 0, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array, the source vector.}

\item{y}{either a cg_node object or a numerical vector or array, the destination vector.}

\item{offset}{a numerical scalar, the offset where \code{x} is copied into \code{y} (optional).}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Efficiently copy all elements of vector \code{x} into \code{y}.
}
\note{
The value of \code{y} is modified in-place. Hence, be cautious that no other node depends on \code{y} as this may yield incorrect gradients when performing a backward pass. Rounding may take place if \code{x} does not match the datatype of \code{y}.
}
\examples{
# Initialize a computational graph
graph <- cg_graph()

# Create a parameter
a <- cg_parameter(1:4)

# Create a constant
b <- cg_constant(rep(0, 10))

# Copy the elements of a into b
c <- cg_copy(a, b, offset = 4)

# The value of b is modified in-place and c holds a shallow copy of b's value
b$value
c$value

}
\author{
Ron Triepels
}
