% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_exp}
\alias{cg_exp}
\title{Exponential Function}
\usage{
cg_exp(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{exp(x)}.
}
\seealso{
\link[base:log]{exp}
}
\author{
Ron Triepels
}
