% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cg_initialize}
\alias{cg_initialize}
\title{Computational Graph}
\arguments{
\item{library}{environment, function library used by the graph. Defaults to \link[cgraph:cg_default_library]{cg_default_library}.}
}
\value{
cgraph object.
}
\description{
Initialize a computational graph.
}
\details{
\code{$new(library = cgraph::cg_default_library)}
}
\note{
The cgraph object is set to be the active graph.
}
\examples{
# Initialize a new computational graph.
x <- cgraph$new()

}
\author{
Ron Triepels
}
