% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{cg_input}
\alias{cg_input}
\title{Add Input}
\usage{
cg_input(name = NULL)
}
\arguments{
\item{name}{character scalar, name of the node (optional). In case argument \code{name} is missing, the node is added to the graph under an automatically generated name.}
}
\value{
cg_node object.
}
\description{
Add an input node to the active graph.
}
\note{
Argument \code{name} is deprecated and will be removed in the next major release.

Inputs cannot be assigned a fixed value but behave as placeholders. Values can be assigned to inputs when evaluating or differentiating a graph (see \link[cgraph]{cg_graph_run} and \link[cgraph]{cg_graph_gradients} for more details).
}
\examples{
# Initialize a computational graph
graph <- cg_graph()

# Add an input to the graph.
a <- cg_input(name = "a")

}
\author{
Ron Triepels
}
