% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_length}
\alias{cg_length}
\title{Length of an Object}
\usage{
cg_length(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{length(x)}.
}
\note{
This operator is not differentiable. Any attempt to differentiate this operator will result in an error.
}
\seealso{
\link[base]{length}
}
\author{
Ron Triepels
}
