% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_linear}
\alias{cg_linear}
\title{Linear Transformation}
\usage{
cg_linear(x, y, z, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric matrix.}

\item{y}{cg.node, placeholder for a numeric matrix.}

\item{z}{cg.node, placeholder for a numeric vector.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{linear(x, y, z)}.
}
\note{
This function is equivalent to \code{cg.matmul(x, y) + cg.as.numeric(z)}.
}
\seealso{
\link[cgraph:linear]{linear}
}
\author{
Ron Triepels
}
