% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_linear}
\alias{cg_linear}
\title{Linear Transformation}
\usage{
cg_linear(x, y, z, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix.}

\item{y}{either a cg_node object or a numerical matrix.}

\item{z}{either a cg_node object or a numerical vector.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{x \%*\% y + c(z)}.
}
\note{
This function is equivalent to \code{cg_matmul(x, y) + cg_as_numeric(z)}.
}
\author{
Ron Triepels
}
