% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_log10}
\alias{cg_log10}
\title{Logarithm Base 10}
\usage{
cg_log10(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{log10(x)}.
}
\seealso{
\link[base:log]{log10}
}
\author{
Ron Triepels
}
