% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_log2}
\alias{cg_log2}
\title{Logarithm Base 2}
\usage{
cg_log2(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{log2(x)}.
}
\seealso{
\link[base:log]{log}
}
\author{
Ron Triepels
}
