% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_max}
\alias{cg_max}
\title{Maxima}
\usage{
cg_max(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{max(x)}.
}
\note{
Function \link[base:Extremes]{max} is called without changing the default value of argument \code{na.rm}.
}
\seealso{
\link[base:Extremes]{max}
}
\author{
Ron Triepels
}
