% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_mean}
\alias{cg_mean}
\title{Arithmetic Mean}
\usage{
cg_mean(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{mean(x)}.
}
\note{
Function \link[base:mean]{mean} is called without setting argument \code{trim} and \code{na.rm}.
}
\seealso{
\link[base:mean]{mean}
}
\author{
Ron Triepels
}
