% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_min}
\alias{cg_min}
\title{Minima}
\usage{
cg_min(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{min(x)}.
}
\note{
Function \link[base:Extremes]{min} is called without changing the default value of argument \code{na.rm}.
}
\seealso{
\link[base:Extremes]{min}
}
\author{
Ron Triepels
}
