% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_mul}
\alias{cg_mul}
\title{Multiply}
\usage{
cg_mul(x, y, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numeric vector or array.}

\item{y}{either a cg_node object or a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{x * y}.
}
\seealso{
\link[base:Arithmetic]{multiply}
}
\author{
Ron Triepels
}
