% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_neg}
\alias{cg_neg}
\title{Negative}
\usage{
cg_neg(n1, name = NULL)
}
\arguments{
\item{name}{character scalar, name of the operation (optional).}

\item{x}{either a cg_node object or a numeric vector or array.}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{-x}.
}
\seealso{
\link[base:Arithmetic]{negative}
}
\author{
Ron Triepels
}
