% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_nrow}
\alias{cg_nrow}
\title{Number of Rows of an Array}
\usage{
cg_nrow(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{nrow(x)}.
}
\note{
This operator is not differentiable. Any attempt to differentiate this operator will result in an error.
}
\seealso{
\link[base]{nrow}
}
\author{
Ron Triepels
}
