% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{cg_parameter}
\alias{cg_parameter}
\title{Add Parameter}
\usage{
cg_parameter(value, name = NULL)
}
\arguments{
\item{value}{numerical vector or array, value of the node.}

\item{name}{character scalar, name of the node (optional). In case argument \code{name} is missing, the node is added to the graph under an automatically generated name.}
}
\value{
cg_node object.
}
\description{
Add a parameter to the active graph.
}
\note{
Parameters are assumed to be subject to some optimization process. Hence, their value might change over time. You can use data member \code{value} of a cg_node object to retrieve or change its value.
}
\examples{
# Initialize a computational graph
graph <- cg_graph()

# Add a parameter with value 1 and name 'a' to the graph.
a <- cg_parameter(1, name = "a")

}
\author{
Ron Triepels
}
