% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_pmin}
\alias{cg_pmin}
\title{Parallel Minima}
\usage{
cg_pmin(x, y, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{y}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{pmin(x, y)}.
}
\note{
Function \link[base:Extremes]{pmin} is called without changing the default value of argument \code{na.rm}.
}
\seealso{
\link[base:Extremes]{pmin}
}
\author{
Ron Triepels
}
