% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_pos}
\alias{cg_pos}
\title{Positive}
\usage{
cg_pos(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x}.
}
\seealso{
\link[base:Arithmetic]{positive}
}
\author{
Ron Triepels
}
