% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_pow}
\alias{cg_pow}
\title{Power}
\usage{
cg_pow(x, y, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{y}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{x ^ y}.
}
\seealso{
\link[base:Arithmetic]{power}
}
\author{
Ron Triepels
}
