% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_rowmeans}
\alias{cg_rowmeans}
\title{Row Means}
\usage{
cg_rowmeans(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{rowMeans(x)}.
}
\note{
Function \link[base:colSums]{rowMeans} is called without changing the default value of argument \code{na.rm} and \code{dims}.
}
\seealso{
\link[base:colSums]{rowMeans}
}
\author{
Ron Triepels
}
