% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_rowsums}
\alias{cg_rowsums}
\title{Row Sums}
\usage{
cg_rowsums(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numeric matrix or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{rowSums(x)}.
}
\note{
Function \link[base:colSums]{rowSums} is called without changing the default value of argument \code{na.rm} and \code{dims}.
}
\seealso{
\link[base:colSums]{rowSums}
}
\author{
Ron Triepels
}
