% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cg_sin}
\alias{cg_sin}
\title{Sine}
\usage{
cg_sin(x, name = NULL)
}
\arguments{
\item{x}{cg.node, placeholder for a numeric vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg.node, node of the operation.
}
\description{
Calculate \code{sin(x)}.
}
\seealso{
\link[base:Trig]{sin}
}
\author{
Ron Triepels
}
