% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cg_sqrt}
\alias{cg_sqrt}
\title{Square Root}
\usage{
cg_sqrt(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object.
}
\description{
Calculate \code{sqrt(x)}.
}
\seealso{
\link[base:MathFun]{sqrt}
}
\author{
Ron Triepels
}
