% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_sub}
\alias{cg_sub}
\title{Subtract}
\usage{
cg_sub(n1, n2, name = NULL)
}
\arguments{
\item{name}{character scalar, name of the operation (optional).}

\item{x}{either a cg_node object or a numeric vector or array.}

\item{y}{either a cg_node object or a numeric vector or array.}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{x - y}.
}
\seealso{
\link[base:Arithmetic]{subtract}
}
\author{
Ron Triepels
}
