% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{cg_t}
\alias{cg_t}
\title{Matrix Transpose}
\usage{
cg_t(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical matrix.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{t(x)}.
}
\seealso{
\link[base:t]{t}
}
\author{
Ron Triepels
}
