% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cg_tan}
\alias{cg_tan}
\title{Tangent}
\usage{
cg_tan(x, name = NULL)
}
\arguments{
\item{x}{either a cg_node object or a numerical vector or array.}

\item{name}{character scalar, name of the operation (optional).}
}
\value{
cg_operator object, node of the operation.
}
\description{
Calculate \code{tan(x)}.
}
\seealso{
\link[base:Trig]{tan}
}
\author{
Ron Triepels
}
