% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cgraph}
\alias{cgraph}
\title{Computational Graph}
\description{
The \code{cgraph} class facilitates the construction, evaluation, and differentiation of computaiontal graphs in R.
}
\section{Usage}{

\preformatted{x <- cgraph$new()}
}

\section{Members}{

\describe{
\item{x$nodes}{list, symbolic representation of the nodes.}
\item{x$values}{environment, values of the nodes.}
}
}

\section{Methods}{

\describe{
\item{x$initialize}{initialize a computational graph, see \link[cgraph]{cg.initialize}.}
\item{x$name}{generate a default name for a node, see \link[cgraph]{cg.name}.}
\item{x$const}{add a constant node to the graph, see \link[cgraph]{cg.const}.}
\item{x$input}{add an input node to the graph, see \link[cgraph]{cg.input}.}
\item{x$parm}{add a parameter node to the graph, see \link[cgraph]{cg.parm}.}
\item{x$get.parms}{list all parameters and their values, see \link[cgraph]{cg.get.parms}.}
\item{x$add.parms}{add parameters to the graph, see \link[cgraph]{cg.add.parms}.}
\item{x$opr}{add an operation node to the graph, see \link[cgraph]{cg.opr}.}
\item{x$active}{set the graph to be the active graph, see \link[cgraph]{cg.active}.}
\item{x$run}{evaluate a node in the graph, see \link[cgraph]{cg.run}.}
\item{x$gradients}{differentiate the graph by reverse automatic differentiation, see \link[cgraph]{cg.gradients}.}
\item{x$approx.grad}{differentiate the graph by numerical differentiation, see \link[cgraph]{cg.approx.grad}.}
\item{x$adj.mat}{retrieve the adjacency matrix of the graph, see \link[cgraph]{cg.adj.mat}.}
\item{x$plot}{plot the topology of the graph, see \link[cgraph]{cg.plot}.}
}
}

\author{
Ron Triepels
}
