% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{const}
\alias{const}
\title{Add Constant}
\value{
symbol, the name of the node.
}
\description{
Add a constant node to the graph.
}
\note{
Constant nodes are ignored when differentiating a graph.
}
\section{Usage}{

\preformatted{const(value, name)}
}

\section{Agruments}{

\describe{
\item{value}{numeric scalar or array, default value of the node.}
\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, an auto-generated name is assigned to the node.}
}
}

\author{
Ron Triepels
}
