% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{const}
\alias{const}
\title{Add Constant}
\usage{
const(value, name)
}
\arguments{
\item{value}{numeric scalar or array, default value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, an auto-generated name is assigned to the node.}
}
\value{
cg.node, constant node.
}
\description{
Add a constant node to the active graph.
}
\note{
Constant nodes are ignored when differentiating a graph.
}
\author{
Ron Triepels
}
