% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{initialize}
\alias{initialize}
\title{Computational Graph}
\value{
cgraph object.
}
\description{
Initialize a new computational graph.
}
\note{
Functions can be made available to a graph by making their environment the enclosure of the graph via argument \code{parent}. The enclosure of a graph should not be another \code{cgraph} object.
}
\section{Usage}{

\preformatted{cgraph$new(parent)}
}

\section{Agruments}{

\describe{
\item{parent}{environment, the environment to be used as the enclosure of the graph created.}
}
}

\author{
Ron Triepels
}
