% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{input}
\alias{input}
\title{Add Input}
\usage{
input(value, name)
}
\arguments{
\item{value}{numeric scalar or array, default value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, the node is tried to be added to the graph under an auto-generated name.}
}
\value{
cg.node, input node.
}
\description{
Add an input node to the active graph.
}
\note{
The name of the node cannot be \code{grad} as this is a reserved word.
}
\author{
Ron Triepels
}
