% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{input}
\alias{input}
\title{Add Input}
\usage{
input(value, name)
}
\arguments{
\item{value}{numeric scalar or array, default value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, an auto-generated name is assigned to the node.}
}
\value{
cg.node, input node.
}
\description{
Add an input node to the current graph.
}
\author{
Ron Triepels
}
