% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{linear}
\alias{linear}
\title{Linear Transformation}
\usage{
linear(x, y, z)
}
\arguments{
\item{x}{numeric matrix.}

\item{y}{numeric matrix.}

\item{z}{numeric vector.}
}
\value{
numeric matrix, result of the transformation.
}
\description{
Calculate the linear transformation \code{x \%*\% y + c(z)}.
}
\author{
Ron Triepels
}
