% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{linear}
\alias{linear}
\title{Linear Transformation}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{x \%*\% y + c(z)}.
}
\note{
This function is equivalent to \code{add(mat.mul(x, y), c(z))}.
}
\section{Usage}{

\preformatted{linear(x, y, z, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric matrix.}
\item{y}{character scalar or symbol, placeholder for a numeric matrix.}
\item{z}{character scalar or symbol, placeholder for a numeric vector.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
