% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{name}
\alias{name}
\title{Generate Name}
\usage{
name()
}
\value{
character scalar, auto-generated name for the node.
}
\description{
Generate a default name for a node.
}
\note{
The auto-generated name is not guaranteed to be unique.
}
\examples{
# Initialize a new computational graph.
x <- cgraph$new()

# Generate a name.
name()

}
\author{
Ron Triepels
}
