% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{parm}
\alias{parm}
\title{Add Parameter}
\usage{
parm(value, name)
}
\arguments{
\item{value}{numeric scalar or array, default value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, an auto-generated name is assigned to the node.}
}
\value{
cg.node, parameter node.
}
\description{
Add a parameter node to the current graph.
}
\note{
In case \code{value} has no dimension (i.e. it has no attribute \code{dim} attached to it), the dimension of \code{value} is automatically set to the length of the vector.
}
\author{
Ron Triepels
}
