% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{parm}
\alias{parm}
\title{Add Parameter}
\usage{
parm(value, name)
}
\arguments{
\item{value}{numeric scalar or array, default value of the node.}

\item{name}{character scalar or symbol, name of the node (optional). In case \code{name} is missing, the node is tried to be added to the graph under an auto-generated name.}
}
\value{
cg.node, parameter node.
}
\description{
Add a parameter node to the active graph.
}
\note{
The name of the node cannot be \code{grad} as this is a reserved word.
}
\author{
Ron Triepels
}
