% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{reshape_grad_x}
\alias{reshape_grad_x}
\title{Reshape Array Dimensions Gradient}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{dim}{numeric scalar or vector, value of \code{dim}.}

\item{val}{numeric vector or array, value of \code{array(x)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{array(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
