% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run}
\alias{run}
\title{Evaluate a Graph}
\usage{
run(name, values = list())
}
\arguments{
\item{name}{character scalar or symbol, name of the node that needs to be evaluated.}

\item{values}{named list or environment, values that are subsituted for the placeholders in the graph.}
}
\value{
cg.results object, the value of node \code{name} including the values of all ancestors of node \code{name} that are evaluated in the forward-pass.
}
\description{
Evaluate node \code{name} in the current graph.
}
\details{
\code{$run(name, values = list())}
}
\note{
All placeholders required to compute node \code{name} must have a value. Placeholders can be assigned a default value when they are created. Alternatively, argument \code{values} can be used to substitute values for placeholders that do not have a default value or to fix the values of nodes.

Only those nodes needed to compute node \code{name} are evaluated and their values are returned. The values of placeholders whose default values are not changed are not returned.
}
\author{
Ron Triepels
}
