% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{set}
\alias{set}
\title{Change the Default Value of a Node}
\usage{
set(name, value)
}
\arguments{
\item{name}{cg.node, name of the node whose default value is to be changed.}

\item{value}{numeric vector or array, default value of the node.}
}
\value{
nothing.
}
\description{
Change the default value of a node in the active graph.
}
\author{
Ron Triepels
}
