% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid}
\value{
symbol, name of the operation.
}
\description{
Calculate \code{1 / (1 + exp(1)^-x)}.
}
\note{
\code{exp(1)=e}.
}
\section{Usage}{

\preformatted{sigmoid(x, name)}
}

\section{Agruments}{

\describe{
\item{x}{character scalar or symbol, placeholder for a numeric scalar or array.}
\item{name}{character scalar, name of the operation (optional).}
}
}

\author{
Ron Triepels
}
