% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sigmoid_grad}
\alias{sigmoid_grad}
\title{Sigmoid Gradient}
\usage{
sigmoid_grad(x, val, grad)
}
\arguments{
\item{grad}{numeric vector or array, gradient of \code{x}.}

\item{y}{numeric vector or array, value of \code{1 / (1 + exp(-x))}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{1 / (1 + exp(-x))} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
