% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{sum_grad}
\alias{sum_grad}
\title{Sum of Vector Elements Gradient}
\arguments{
\item{x}{numeric vector or array, value of \code{x}.}

\item{val}{numeric vector or array, value of \code{sum(x)}.}

\item{grad}{numeric vector or array, gradient of \code{x}.}
}
\value{
numeric vector or array, gradient of the operation.
}
\description{
Calculate the gradient of \code{sum(x)} with respect to \code{x}.
}
\author{
Ron Triepels
}
\keyword{internal}
